# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/planet": {
            "get": {
                "description": "Возращает список всех активных планетаей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Получение списка планеты",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Код планеты",
                        "name": "searchCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    }
                }
            }
        },
        "/planet/create": {
            "post": {
                "description": "Создает новый планета с предоставленными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Создание нового планеты",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Код планеты",
                        "name": "searchCode",
                        "in": "query"
                    },
                    {
                        "description": "Пользовательский объект в формате JSON",
                        "name": "planet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PlanetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    }
                }
            }
        },
        "/planet/{planet_id}": {
            "get": {
                "description": "Возвращает информацию о планетае по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Получение планеты по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о планетае",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    }
                }
            }
        },
        "/planet/{planet_id}/delete": {
            "delete": {
                "description": "Удаляет планета по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Удаление планеты",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Код планеты",
                        "name": "searchCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    }
                }
            }
        },
        "/planet/{planet_id}/constellation": {
            "post": {
                "description": "Добавляет планета к доставке по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Добавление планеты к доставке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Код планеты",
                        "name": "searchCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    }
                }
            }
        },
        "/planet/{planet_id}/image": {
            "post": {
                "description": "Добавляет изображение к планетау по его ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Добавление изображения к планетау",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение планеты",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    }
                }
            }
        },
        "/planet/{planet_id}/update": {
            "put": {
                "description": "Обновляет информацию о планетае по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Обновление информации о планетае",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о планетае",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Planet"
                        }
                    }
                }
            }
        },
        "/planets/{planet_id}/constellation": {
            "post": {
                "description": "Удаляет планета из созвездия по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Планета"
                ],
                "summary": "Удаление планеты из созвездия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID планеты",
                        "name": "planet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Код планеты",
                        "name": "searchCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список планетаей",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PlanetsGetResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Planet": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string",
                    "example": "AirlineX"
                },
                "planet_code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "planet_id": {
                    "type": "integer"
                },
                "planet_status": {
                    "type": "string",
                    "example": "checked"
                },
                "planet_type": {
                    "type": "string",
                    "example": "suitcase"
                },
                "owner_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "pasport_details": {
                    "type": "string",
                    "example": "123456789"
                },
                "photo": {
                    "type": "string",
                    "example": "http://example.com/planet.jpg"
                },
                "size": {
                    "type": "string",
                    "example": "large"
                },
                "weight": {
                    "type": "number",
                    "example": 23.5
                }
            }
        },
        "model.PlanetRequest": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string",
                    "example": "AirlineX"
                },
                "planet_code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "planet_type": {
                    "type": "string",
                    "example": "suitcase"
                },
                "owner_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "pasport_details": {
                    "type": "string",
                    "example": "123456789"
                },
                "size": {
                    "type": "string",
                    "example": "large"
                },
                "weight": {
                    "type": "number",
                    "example": 23.5
                }
            }
        },
        "model.PlanetsGetResponse": {
            "type": "object",
            "properties": {
                "planets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Planet"
                    }
                },
                "constellation_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}